Keep navigation handler tests easier to extend
The navigation handler test suite repeated DOM mock setup and element builders inline, so this extracts the shared mock environment helpers into a sibling module without changing click, hover, or scroll assertions.

Constraint: Refactor must preserve navigation handler test behavior and timing-sensitive hover assertions exactly
Constraint: veryfront-code pre-commit verify:quick currently fails on pre-existing CLI boundary violations in cli/commands/extension/*.ts
Rejected: Factor callback spies into another helper layer | would widen this pass beyond the narrow DOM mock extraction targeted by the wave
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep reusable navigation handler DOM test setup in the helper module unless a future case needs bespoke inline environment state
Tested: deno test --no-check --allow-all src/routing/client/navigation-handlers.test.ts; deno fmt --check src/routing/client/navigation-handlers.test.ts src/routing/client/navigation-handlers.test-helpers.ts; deno lint src/routing/client/navigation-handlers.test.ts src/routing/client/navigation-handlers.test-helpers.ts; git diff --check
Not-tested: Full veryfront-code repo suite
